introduce helpers_*.go files
