+++ internal/checkers/float_compare.go
@@ -2,9 +2,7 @@ package checkers
 
 import (
 	"fmt"
-	"go/ast"
 	"go/token"
-	"go/types"
 
 	"golang.org/x/tools/go/analysis"
 )
@@ -33,10 +31,10 @@ func (checker FloatCompare) Check(pass *analysis.Pass, call *CallMeta) *analysis
 			return len(call.Args) > 1 && (isFloat(pass, call.Args[0]) || isFloat(pass, call.Args[1]))
 
 		case "True":
-			return len(call.Args) > 0 && isFloatCompare(pass, call.Args[0], token.EQL)
+			return len(call.Args) > 0 && isComparisonWithFloat(pass, call.Args[0], token.EQL)
 
 		case "False":
-			return len(call.Args) > 0 && isFloatCompare(pass, call.Args[0], token.NEQ)
+			return len(call.Args) > 0 && isComparisonWithFloat(pass, call.Args[0], token.NEQ)
 		}
 		return false
 	}()
@@ -49,22 +47,4 @@ func (checker FloatCompare) Check(pass *analysis.Pass, call *CallMeta) *analysis
 		return newDiagnostic(checker.Name(), call, fmt.Sprintf(format, call.SelectorXStr), nil)
 	}
 	return nil
-}
-
-func isFloat(pass *analysis.Pass, expr ast.Expr) bool {
-	t := pass.TypesInfo.TypeOf(expr)
-	if t == nil {
-		return false
-	}
-
-	bt, ok := t.Underlying().(*types.Basic)
-	return ok && (bt.Info()&types.IsFloat > 0)
-}
-
-func isFloatCompare(p *analysis.Pass, e ast.Expr, op token.Token) bool {
-	be, ok := e.(*ast.BinaryExpr)
-	if !ok {
-		return false
-	}
-	return be.Op == op && (isFloat(p, be.X) || isFloat(p, be.Y))
 }
