+++ internal/checkers/helpers_diagnostic.go
