+++ internal/checkers/helpers_http.go
@@ -1,0 +1,35 @@
+package checkers
+
+import (
+	"go/ast"
+	"go/types"
+
+	"golang.org/x/tools/go/analysis"
+
+	"github.com/Antonboom/testifylint/internal/analysisutil"
+)
+
+func mimicHTTPHandler(pass *analysis.Pass, fType *ast.FuncType) bool {
+	httpHandlerFuncObj := analysisutil.ObjectOf(pass.Pkg, "net/http", "HandlerFunc")
+	if httpHandlerFuncObj == nil {
+		return false
+	}
+
+	sig, ok := httpHandlerFuncObj.Type().Underlying().(*types.Signature)
+	if !ok {
+		return false
+	}
+
+	if len(fType.Params.List) != sig.Params().Len() {
+		return false
+	}
+
+	for i := 0; i < sig.Params().Len(); i++ {
+		lhs := sig.Params().At(i).Type()
+		rhs := pass.TypesInfo.TypeOf(fType.Params.List[i].Type)
+		if !types.Identical(lhs, rhs) {
+			return false
+		}
+	}
+	return true
+}
