+++ internal/checkers/helpers_naming.go
@@ -1,0 +1,21 @@
+package checkers
+
+import (
+	"go/ast"
+	"regexp"
+)
+
+func isIdentNamedAsExpected(pattern *regexp.Regexp, e ast.Expr) bool {
+	id, ok := e.(*ast.Ident)
+	return ok && pattern.MatchString(id.Name)
+}
+
+func isStructVarNamedAsExpected(pattern *regexp.Regexp, e ast.Expr) bool {
+	s, ok := e.(*ast.SelectorExpr)
+	return ok && isIdentNamedAsExpected(pattern, s.X)
+}
+
+func isStructFieldNamedAsExpected(pattern *regexp.Regexp, e ast.Expr) bool {
+	s, ok := e.(*ast.SelectorExpr)
+	return ok && isIdentNamedAsExpected(pattern, s.Sel)
+}
