+++ internal/checkers/helpers_nil.go
@@ -1,0 +1,19 @@
+package checkers
+
+import "go/ast"
+
+func xorNil(first, second ast.Expr) (ast.Expr, bool) {
+	a, b := isNil(first), isNil(second)
+	if xor(a, b) {
+		if a {
+			return second, true
+		}
+		return first, true
+	}
+	return nil, false
+}
+
+func isNil(expr ast.Expr) bool {
+	ident, ok := expr.(*ast.Ident)
+	return ok && ident.Name == "nil"
+}
