+++ internal/checkers/nil_compare.go
@@ -1,8 +1,6 @@
 package checkers
 
 import (
-	"go/ast"
-
 	"golang.org/x/tools/go/analysis"
 
 	"github.com/Antonboom/testifylint/internal/analysisutil"
@@ -55,15 +53,4 @@ func (checker NilCompare) Check(pass *analysis.Pass, call *CallMeta) *analysis.D
 			NewText: analysisutil.NodeBytes(pass.Fset, survivingArg),
 		}),
 	)
-}
-
-func xorNil(first, second ast.Expr) (ast.Expr, bool) {
-	a, b := isNil(first), isNil(second)
-	if xor(a, b) {
-		if a {
-			return second, true
-		}
-		return first, true
-	}
-	return nil, false
 }
