+++ internal/checkers/suite_thelper.go
@@ -3,13 +3,11 @@ package checkers
 import (
 	"fmt"
 	"go/ast"
-	"strings"
 
 	"golang.org/x/tools/go/analysis"
 	"golang.org/x/tools/go/ast/inspector"
 
 	"github.com/Antonboom/testifylint/internal/analysisutil"
-	"github.com/Antonboom/testifylint/internal/testify"
 )
 
 // SuiteTHelper requires t.Helper() call in suite helpers:
@@ -27,15 +25,15 @@ func (SuiteTHelper) Name() string   { return "suite-thelper" }
 func (checker SuiteTHelper) Check(pass *analysis.Pass, inspector *inspector.Inspector) (diagnostics []analysis.Diagnostic) {
 	inspector.Preorder([]ast.Node{(*ast.FuncDecl)(nil)}, func(node ast.Node) {
 		fd := node.(*ast.FuncDecl)
-		if !isTestifySuiteMethod(pass, fd) {
+		if !isSuiteMethod(pass, fd) {
 			return
 		}
 
-		if ident := fd.Name; ident == nil || isTestMethod(ident.Name) || isServiceMethod(ident.Name) {
+		if ident := fd.Name; ident == nil || isSuiteTestMethod(ident.Name) || isSuiteServiceMethod(ident.Name) {
 			return
 		}
 
-		if !containsSuiteAssertions(pass, fd) {
+		if !containsAssertions(pass, fd) {
 			return
 		}
 
@@ -68,43 +66,20 @@ func (checker SuiteTHelper) Check(pass *analysis.Pass, inspector *inspector.Insp
 	return diagnostics
 }
 
-func isTestifySuiteMethod(pass *analysis.Pass, fDecl *ast.FuncDecl) bool {
-	if fDecl.Recv == nil || len(fDecl.Recv.List) != 1 {
-		return false
-	}
-
-	rcv := fDecl.Recv.List[0]
-	return implementsTestifySuite(pass, rcv.Type)
-}
-
-func isTestMethod(name string) bool {
-	return strings.HasPrefix(name, "Test")
-}
-
-func isServiceMethod(name string) bool {
-	// https://github.com/stretchr/testify/blob/master/suite/interfaces.go
-	switch name {
-	case "T", "SetT", "SetS", "SetupSuite", "SetupTest", "TearDownSuite", "TearDownTest",
-		"BeforeTest", "AfterTest", "HandleStats", "SetupSubTest", "TearDownSubTest":
-		return true
-	}
-	return false
-}
-
-func containsSuiteAssertions(pass *analysis.Pass, fn *ast.FuncDecl) bool {
+func containsAssertions(pass *analysis.Pass, fn *ast.FuncDecl) bool {
 	if fn.Body == nil {
 		return false
 	}
 
 	for _, s := range fn.Body.List {
-		if isSuiteAssertion(pass, s) {
+		if isAssertionStmt(pass, s) {
 			return true
 		}
 	}
 	return false
 }
 
-func isSuiteAssertion(pass *analysis.Pass, stmt ast.Stmt) bool {
+func isAssertionStmt(pass *analysis.Pass, stmt ast.Stmt) bool {
 	expr, ok := stmt.(*ast.ExprStmt)
 	if !ok {
 		return false
@@ -115,16 +90,5 @@ func isSuiteAssertion(pass *analysis.Pass, stmt ast.Stmt) bool {
 		return false
 	}
 
-	se, ok := ce.Fun.(*ast.SelectorExpr)
-	if !ok || se.Sel == nil {
-		return false
-	}
-
-	if sel, ok := pass.TypesInfo.Selections[se]; ok {
-		pkg := sel.Obj().Pkg()
-		isAssert := analysisutil.IsPkg(pkg, testify.AssertPkgName, testify.AssertPkgPath)
-		isRequire := analysisutil.IsPkg(pkg, testify.RequirePkgName, testify.RequirePkgPath)
-		return isAssert || isRequire
-	}
-	return false
+	return NewCallMeta(pass, ce) != nil
 }
